feat: useMutation refactor
Refactored code to make declarative fetchs using useMutation
